Named modules, rename Recipe Correctly positioned Settings button, but the settings don't stay open
