+++ vue/components/RecipeImages.vue
@@ -6,6 +6,7 @@
 
 <script>
 export default {
+    name: 'RecipeImages',
     data () {
         return {
             collapsed: true,
