+++ vue/components/RecipeIngredient.vue
@@ -4,7 +4,7 @@
 
 <script>
 export default {
-
+    name: 'RecipeIngredient',
     props: ['ingredient'],
     data () {
         return {
