+++ vue/components/RecipeInstruction.vue
@@ -4,7 +4,7 @@
 
 <script>
 export default {
-
+    name: 'RecipeInstruction',
     props: ['instruction'],
     data () {
         return {
