+++ vue/components/RecipeTool.vue
@@ -4,7 +4,7 @@
 
 <script>
 export default {
-
+    name: 'RecipeTool',
     props: ['tool'],
 
 }
